🫦

+++ WO/mw1/src/s1/mod.rs
@@ -66,7 +66,7 @@ impl Deck {
     }
 
     // Function to deal out random cards
-    fn deal(&mut self, num_cards: usize) {}
+    fn deal(&mut self, num_cards: i8) {}
 }
 
 fn s1_fn() {
